requirements.txt
requests~=2.31.0
pytest~=7.4.3
pydantic~=2.5.3
allure-pytest==2.13.2